flutter_earth/requirements.txt
numpy>=1.21.0
rasterio>=1.2.0
pyshp>=2.1.0
folium>=0.12.0
pywebview>=3.6.0
requests>=2.26.0
python-dateutil>=2.8.2
PyQt6-Qt6>=6.4.0
PyQt6-QtQml>=6.4.0
earthengine-api>=0.1.290
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
scipy>=1.10.1
Pillow>=9.5.0
shapely>=2.0.1
matplotlib>=3.7.1
qtpy>=2.3.1
overpass>=0.7
urllib3>=2.0.3 